DEV: Initializing OmicCircos (2) as RStudio project.
This OmicCircos is forked from the original project, so bumping version out of range. Could also renaome, but don't want to make it too hard to merge back with the original authors, so just soft-naming this (2), formal name is still OmicCircos. Adding my name as author and maintainer of THIS fork.